make it easier to look at the GlobalConsumer code
